Separated out the child and shared struct code into different files for organization. Monolithic code is bad.
