+++ Child.h
@@ -1,0 +1,14 @@
+/*
+ * child.h
+ *
+ *  Created on: Oct 7, 2018
+ *      Author: Gr34v
+ */
+
+#ifndef CHILD_H_
+#define CHILD_H_
+
+
+void child(void *params);
+
+#endif /* CHILD_H_ */
